+++ src/model/user/user.cpp
@@ -7,10 +7,10 @@ User::User(): UserBase() {
     init();
 }
 
-User::User(int id, std::string name, std::string password, int point)
+User::User(int id, std::string name, std::string password, long long point)
     : UserBase(id, std::move(name), std::move(password)), point(point) {}
 
-User::User(std::string name, std::string password, int point)
+User::User(std::string name, std::string password, long long point)
     : UserBase(std::move(name), std::move(password)), point(point) {}
 
 
@@ -25,7 +25,7 @@ bool User::isGuest() {
 }
 
 
-bool User::setPoint(int point) {
+bool User::setPoint(long long point) {
     if (isGuest()) {
         this->point = 0;
         return false;
@@ -35,13 +35,13 @@ bool User::setPoint(int point) {
 }
 
 
-bool User::addPoint(int point) {
+bool User::addPoint(long long point) {
     if (isGuest()) {return false;}
     this->point += point;
     return true;
 }
 
-bool User::usePoint(int point) {
+bool User::usePoint(long long point) {
     if (isGuest() || point < 0 || this->point < point) {
         return false;
     }
@@ -49,7 +49,7 @@ bool User::usePoint(int point) {
     return true;
 }
 
-bool User::subPoint(int point) {
+bool User::subPoint(long long point) {
     if (isGuest() || point < 0) {
         return false;
     }
